Moving method in Query
